Move ActiveFedora.class_from_string into ActiveFedora::Model
