+++ spec/unit/model_spec.rb
@@ -36,4 +36,30 @@ describe ActiveFedora::Model do
       it { should be_nil }
     end
   end
+
+  describe ".class_from_string" do
+    before do
+      module ParentClass
+        class SiblingClass
+        end
+        class OtherSiblingClass
+        end
+      end
+    end
+    it "returns class constants based on strings" do
+      expect(described_class.class_from_string("Om")).to eq Om
+      expect(described_class.class_from_string("ActiveFedora::RDF::IndexingService")).to eq ActiveFedora::RDF::IndexingService
+      expect(described_class.class_from_string("IndexingService", ActiveFedora::RDF)).to eq ActiveFedora::RDF::IndexingService
+    end
+
+    it "finds sibling classes" do
+      expect(described_class.class_from_string("SiblingClass", ParentClass::OtherSiblingClass)).to eq ParentClass::SiblingClass
+    end
+
+    it "raises a NameError if the class isn't found" do
+      expect {
+        described_class.class_from_string("FooClass", ParentClass::OtherSiblingClass)
+      }.to raise_error NameError, /uninitialized constant (Object::)?FooClass/
+    end
+  end
 end
